+++ utils/index.js
@@ -31,12 +31,12 @@ module.exports = (...args) => {
       library.utils["powershell"] = require("./powershell")(params, obj);
       library.utils["intercomm"] = require("./intercomm")(params, obj);
       library.utils["sqlitesession"] = require("./sqlitesession");
+      library.utils["smfetch"] = require("./smfetch")(params, obj);
       library.utils["startupinit"] = require("./startupinit")(params, obj);
       library.utils["excluded"] = [
         "powershell",
         "intercomm",
         "sqlitesession",
-        "html",
         "io",
         "startupinit",
         "excluded",